Switching to iterator and remove thread pool
Not sure, the thread pool executor, the way I was
using it might have been not working correctly

**
This version matches the golden master generated
from old-version branch.

old-version is at c7d1dd91cad066fc1974c7ef5cf4a016c57a9b76

The only difference was that nodes were in a
different order. In old-version they seemed to be
not sorted, but in new they are.
**